src/splitter.py
from langchain_text_splitters import RecursiveCharacterTextSplitter
from .config import CHUNK_SIZE, CHUNK_OVERLAP


def split_documents(docs):
    """
    Разбивает документы на чанки.
    """
    splitter = RecursiveCharacterTextSplitter(
        chunk_size=CHUNK_SIZE,
        chunk_overlap=CHUNK_OVERLAP,
    )
    chunks = splitter.split_documents(docs)
    print(f"[SPLITTER] Число чанков после разбиения: {len(chunks)}")
    return chunks
