requirements.txt
langchain-ollama
langchain-chroma
langchain-community
langchain-core
pypdf
chromadb
python-dotenv